move some of hte ipe instances to a separate subpackage
