+++ cabal.project
@@ -69,9 +69,9 @@ source-repository-package
 --     subdir: jsaddle
 --     tag: e034fd40cd07c23ad85f3b059e8e3d8b698e322a
 
-source-repository-package
-    type: git
-    location: https://github.com/noinia/miso
+-- source-repository-package
+--     type: git
+--     location: https://github.com/noinia/miso
 
 -- package miso
 --     flags: +jsaddle
