+++ hgeometry-examples/hgeometry-examples.cabal
@@ -114,9 +114,7 @@ common setup
 
 common miso-setup
   build-depends:
-        miso
-      , jsaddle                   >= 0.9.9.0       && < 0.10
-      , jsaddle-warp              >= 0.9.9.0       && < 0.10
+        miso                   >= 1.9.0.0    && < 2.0
 
 common quickcheck-setup
     ghc-options:
