+++ hgeometry-svg/hgeometry-svg.cabal
@@ -53,10 +53,10 @@ common all-setup
       , mtl                       >= 2.2           && < 3
       , filepath                  >= 1.4.100       && < 2
       , file-io                   >= 0.1           && < 1
-      , text                    >= 2.1.2         && < 3
-      , colour                  >= 2.3.6         && < 3
-      , vinyl                   >= 0.14          && < 1
-      , singletons              >= 3             && < 4
+      , text                      >= 2.1.2         && < 3
+      , colour                    >= 2.3.6         && < 3
+      , vinyl                     >= 0.14          && < 1
+      , singletons                >= 3             && < 4
       , miso                      >= 1.9           && < 2
       , lucid                     >= 2             && < 3
       , lucid-svg                 >= 0.7           && < 1
