+++ hgeometry/doctests.hs
@@ -1,4 +1,6 @@
 {-# OPTIONS_GHC -F -pgmF doctest-driver-gen
+                 -fno-warn-partial-type-signatures
+
                 -optF src
 
                 -optF -XTypeFamilies
