+++ hgeometry/intersection-src/HGeometry/Cone/Intersection.hs
